requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.8
gunicorn==23.0.0
holidays==0.85
mysqlclient==2.2.7
packaging==25.0
pycparser==2.23
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
